GSYE-261: Move grid validation inside CommunityDatasheetValidator
